refactor: move establishing group for mixed conversation to separate module
